requirements.txt
# Due Diligence System - GCP Version
# Install: pip install -r requirements.txt

# GCP
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0

# Vector Database
qdrant-client>=1.7.0

# Search
rank-bm25>=0.2.2
sentence-transformers>=2.2.2

# Utilities
tiktoken>=0.5.2
numpy>=1.24.0
tqdm>=4.66.0
tiktoken==0.5.2

# Orchestration & Local LLM
langgraph>=0.0.10
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1

# API
fastapi>=0.109.0
uvicorn==0.23.2
streamlit
plotly
python-dotenv>=1.0.0
pandas
requests
